+++ lua/gideon/messages/init.lua
@@ -1,0 +1,12 @@
+local history_table = {}
+
+local function push_message_to_history(table, message)
+	table[#table + 1] = {
+		text = message,
+	}
+end
+
+return {
+	history_table = history_table,
+	push_message_to_history = push_message_to_history,
+}
